RatingBar.Value changed to double
Value is now set as a double that is coerced to a multiple of 0.5 within the Min/Max bounds.

The colors are currently hardcoded, this needs to be addressed. Unit tests should also be added for the coercion.

+++ MaterialDesign3.Demo.Wpf/RatingBar.xaml.cs
@@ -7,7 +7,7 @@ namespace MaterialDesign3Demo
     {
         public RatingBar() => InitializeComponent();
 
-        private void BasicRatingBar_ValueChanged(object sender, RoutedPropertyChangedEventArgs<int> e)
+        private void BasicRatingBar_ValueChanged(object sender, RoutedPropertyChangedEventArgs<double> e)
             => Debug.WriteLine($"BasicRatingBar value changed from {e.OldValue} to {e.NewValue}.");
     }
 }
